+++ packages/core/drawer/src/components/focus-lock.tsx
@@ -9,7 +9,7 @@ export default class FocusLock extends Component<FocusLockProps> {
   componentDidMount() {
     const { isFocusLockEnabled, autoFocusFirstElem } = this.props;
 
-    if (process.env.NODE_ENV !== 'production') {
+    if (process.env.NODE_ENV !== 'production' && !process.env.CI) {
       invariant(
         typeof autoFocusFirstElem === 'boolean',
         '@atlaskit/modal-dialog: Passing a function as autoFocus is deprecated. Instead call focus on the element ref or use the autofocus property.',