+++ packages/core/field-radio-group/src/Radio.js
@@ -6,7 +6,7 @@ import Radio from './RadioBase';
 const RadioWithTheme = withTheme(Radio);
 const emptyTheme = {};
 
-if (process.env.NODE_ENV !== 'production') {
+if (process.env.NODE_ENV !== 'production' && !process.env.CI) {
   // eslint-disable-next-line no-console
   console.warn(
     '@atlaskit/field-radio-group has been deprecated. Please use the @atlaskit/radio package instead.',