+++ packages/core/field-text/src/FieldTextStateless.js
@@ -20,7 +20,7 @@ type Props = FieldTextProps & {
   innerRef?: (node: ?HTMLInputElement) => void,
 };
 
-if (process.env.NODE_ENV !== 'production') {
+if (process.env.NODE_ENV !== 'production' && !process.env.CI) {
   // eslint-disable-next-line no-console
   console.warn(
     '@atlaskit/field-text has been deprecated. Please use the @atlaskit/textfield package instead.',