+++ packages/media/media-picker/src/components/dropzone/dropzone.tsx
@@ -15,7 +15,6 @@ import {
   withAnalyticsEvents,
   withAnalyticsContext,
   WithAnalyticsEventProps,
-  AnalyticsEventPayload,
 } from '@atlaskit/analytics-next';
 
 import { FabricChannel } from '@atlaskit/analytics-listeners';
@@ -164,18 +163,12 @@ export class DropzoneBase extends LocalUploadComponentReact<
   private onDrop = (e: DragEvent): void => {
     if (e.dataTransfer && dragContainsFiles(e)) {
       const dataTransfer = e.dataTransfer;
-      const length = this.getDraggedItemsLength(dataTransfer);
+      const fileCount = this.getDraggedItemsLength(dataTransfer);
 
-      this.fireAnalyticsEvent({
-        action: 'droppedInto',
-        actionSubject: 'dropzone',
-        attributes: {
-          fileCount: length,
-        },
-      });
+      this.fireAnalyticsEvent('droppedInto', fileCount);
 
       if (this.props.onDrop) this.props.onDrop();
-      this.emitDragLeave({ length });
+      this.emitDragLeave({ length: fileCount });
     }
   };
 
@@ -184,13 +177,7 @@ export class DropzoneBase extends LocalUploadComponentReact<
       const { onDragEnter } = this.props;
       this.uiActive = true;
 
-      this.fireAnalyticsEvent({
-        action: 'draggedInto',
-        actionSubject: 'dropzone',
-        attributes: {
-          fileCount: payload.length,
-        },
-      });
+      this.fireAnalyticsEvent('draggedInto', payload.length);
 
       if (onDragEnter) onDragEnter(payload);
     }
@@ -207,13 +194,7 @@ export class DropzoneBase extends LocalUploadComponentReact<
         if (!this.uiActive) {
           const { onDragLeave } = this.props;
 
-          this.fireAnalyticsEvent({
-            action: 'draggedOut',
-            actionSubject: 'dropzone',
-            attributes: {
-              fileCount: payload.length,
-            },
-          });
+          this.fireAnalyticsEvent('draggedOut', payload.length);
 
           if (onDragLeave) onDragLeave(payload);
         }
@@ -221,10 +202,17 @@ export class DropzoneBase extends LocalUploadComponentReact<
     }
   }
 
-  private fireAnalyticsEvent(payload: AnalyticsEventPayload): void {
+  private fireAnalyticsEvent(action: string, fileCount: number): void {
     const { createAnalyticsEvent } = this.props;
     if (createAnalyticsEvent) {
-      const analyticsEvent = createAnalyticsEvent(payload);
+      const analyticsEvent = createAnalyticsEvent({
+        eventType: 'ui',
+        actionSubject: 'dropzone',
+        action,
+        attributes: {
+          fileCount,
+        },
+      });
       analyticsEvent.fire(FabricChannel.media);
     }
   }